+++ system/expressions/expression_builder.h
@@ -33,9 +33,9 @@ class ExpressionBuilder {
    */
   virtual ExpressionValueType_e build(const StringSlice &slice,
                                       ExpressionStore *expressionStore,
-                                      int triggerNum,
+                                      uint16_t triggerNum,
                                       ExpressionVariant_v *parseInto,
-                                      int destinationID) = 0;
+                                      uint16_t destinationID) = 0;
 
   /**
    * Destructor
@@ -51,9 +51,10 @@ class ExpressionBuilder {
 class VarExpressionBuilder : public ExpressionBuilder {
  public:
   ExpressionValueType_e build(const StringSlice &slice,
-                              ExpressionStore *expressionStore, int triggerNum,
+                              ExpressionStore *expressionStore,
+                              uint16_t triggerNum,
                               ExpressionVariant_v *parseInto,
-                              int destinationID) override;
+                              uint16_t destinationID) override;
 };
 
 /**
@@ -64,9 +65,10 @@ class VarExpressionBuilder : public ExpressionBuilder {
 class EventExpressionBuilder : public ExpressionBuilder {
  public:
   ExpressionValueType_e build(const StringSlice &slice,
-                              ExpressionStore *expressionStore, int triggerNum,
+                              ExpressionStore *expressionStore,
+                              uint16_t triggerNum,
                               ExpressionVariant_v *parseInto,
-                              int destinationID) override;
+                              uint16_t destinationID) override;
 };
 
 /**
@@ -77,9 +79,10 @@ class EventExpressionBuilder : public ExpressionBuilder {
 class ConstExpressionBuilder : public ExpressionBuilder {
  public:
   ExpressionValueType_e build(const StringSlice &slice,
-                              ExpressionStore *expressionStore, int triggerNum,
+                              ExpressionStore *expressionStore,
+                              uint16_t triggerNum,
                               ExpressionVariant_v *parseInto,
-                              int destinationID) override;
+                              uint16_t destinationID) override;
 };
 
 /**
@@ -93,9 +96,10 @@ class ConstExpressionBuilder : public ExpressionBuilder {
 class UnaryFuncExpressionBuilder : public ExpressionBuilder {
  public:
   ExpressionValueType_e build(const StringSlice &slice,
-                              ExpressionStore *expressionStore, int triggerNum,
+                              ExpressionStore *expressionStore,
+                              uint16_t triggerNum,
                               ExpressionVariant_v *parseInto,
-                              int destinationID) override;
+                              uint16_t destinationID) override;
 };
 
 /**
@@ -109,9 +113,10 @@ class UnaryFuncExpressionBuilder : public ExpressionBuilder {
 class BinaryFuncExpressionBuilder : public ExpressionBuilder {
  public:
   ExpressionValueType_e build(const StringSlice &slice,
-                              ExpressionStore *expressionStore, int triggerNum,
+                              ExpressionStore *expressionStore,
+                              uint16_t triggerNum,
                               ExpressionVariant_v *parseInto,
-                              int destinationID) override;
+                              uint16_t destinationID) override;
 };
 
 #endif  // SYSTEM_EXPRESSIONS_EXPRESSION_BUILDER_H_
