+++ system/states/state_post_flight.h
@@ -5,7 +5,7 @@
 
 class PostFlightState : public FlightState {
  public:
-  PostFlightState(int id, uint32_t period_ms)
+  PostFlightState(uint8_t id, uint32_t period_ms)
       : FlightState(id, period_ms, true, true, true) {}
 
   void init(void) override;
